treat defaults, which may be ENV vars, as user input
By running them through the function again as both the value and the
default value, a 0 will be changed to false. Note that "false"
as an ENV value will be treated as 0, since `to_i` is called.

This is the smallest possible change, but there's a high degree of
coupling between read_timeout_property and initialize. Consider a
rewrite to encapsulate the logic of handling the existence of ENV
settings in read_timeout_property.