Allow sending SIGTERM to workers on timeout
When rack-timeout fires then it can put an application in a "bad" state. This is due to https://www.schneems.com/2017/02/21/the-oldest-bug-in-ruby-why-racktimeout-might-hose-your-server/.

We can give applications the option of restarting the entire process when a timeout is hit by sending it a SIGTERM. Why would we want that?

While the `Thread.raise` api is unsafe, restarting a process is safe. If the application is running a multiple-worker webserver such as unicorn or puma, when the worker process gets a SIGTERM then it will be shut down and restarted by the "master" process.

In puma when `SIGTERM` is sent to a worker, that worker will stop accepting new requests, but continue to process all existing requests currently running in threads. This means that sending a SIGTERM to the current process will not affect requests that are currently running (this is a good thing).

What is the downside of sending a TERM to the current worker when there is a timeout? When you TERM a worker, you will lose some capacity to process requests until puma brings up another process. Puma may take a second or two to boot a new process. If all processes on a dyno are TERM'd then your app would stop serving requests until puma boots new processes which, again, would take time. If the application is hitting timeouts because it is under provisioned then sending TERM to the process, reduces processing power and adds a delay in processing new requests. In the worst case this could mean that the application is stuck in a process restart loop. 

Anecdotally I've seen that usually applications can tolerate a number of timeout events. I say, usually fewer than a dozen a day tend to not experience problems. I don't have an empirical number, but if I had to guess, I would say that it would make sense if it was related to MAX_THREADS. Why?

Imagine that every rack-timeout request that fails corrupts something. It's likely to be a connection of some kind, and if you've got MAX_THREADS=5 then you've got 5 database connections.

Since this is a theory I'm making the setting configurable so we can experiment with it in the wild.