Be able to configure Rack::Timeout with strings
This was originally added as part of #157, and released in 0.6.0.

Then it was reverted in #161, but I think it should be kept in, as
removing it is a breaking change. (Yeah, I know this gem is not 1.0 yet,
but will it ever be? :-) It has soon existed for 10 years.)

It is useful, as it allows you to use Rack::Timeout like this:

    use Rack::Timeout, service_timeout: ENV.fetch("RACK_TIMEOUT_SERVICE_TIMEOUT")