Add upper bound to compatible Rails versions
`Rails::VERSION::MAJOR >= 3` implies this is a future proof gem that will work with `rails ~>12.3.4` whenever it is released in the future. This proposed change errors on the side of caution to better outline which known major versions of `rails` are known to be compatible with this specific version of this gem.